trip-manager/requirements.txt
pymongo==4.3.3
kafka-python==2.0.2
requests==2.28.2
dnspython==2.3.0